+++ StarLevelSystem/Data/RaidsData.cs
@@ -38,9 +38,9 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_gdking" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Greyling", MaxSpawned = 4, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "Neck",    MaxSpawned = 2, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.Alerted },
-                        new RaidSpawnEntry() { PrefabName = "Boar",    MaxSpawned = 2, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.Alerted },
+                        new RaidSpawnEntry() { PrefabName = "Greyling", MaxSpawned = 20, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Neck",    MaxSpawned = 10, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.Alerted },
+                        new RaidSpawnEntry() { PrefabName = "Boar",    MaxSpawned = 10, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.Alerted },
                     },
                 }},
                 { new RaidDefinition() {
@@ -57,7 +57,7 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_bonemass" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Troll", MaxSpawned = 3, SpawnInterval = 30f, SpawnChance = 100f, LevelMin = 1, LevelMax = 3, CreatureAI = AI.AgitatedByBuild, SpawnGroupSize = 2, 
+                        new RaidSpawnEntry() { PrefabName = "Troll", MaxSpawned = 9, SpawnInterval = 30f, SpawnChance = 100f, LevelMin = 1, LevelMax = 3, CreatureAI = AI.AgitatedByBuild, SpawnGroupSize = 2, 
                             CustomCreatureLevelUpChance = new SortedDictionary<int, float>() {
                                 { 1, 50f },
                                 { 2, 25f },
@@ -83,9 +83,9 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_bonemass" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Greydwarf",        MaxSpawned = 6, SpawnInterval = 3f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3},
-                        new RaidSpawnEntry() { PrefabName = "Greydwarf_Elite",  MaxSpawned = 2, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.AgitatedByBuild },
-                        new RaidSpawnEntry() { PrefabName = "Greydwarf_Shaman", MaxSpawned = 2, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.Alerted },
+                        new RaidSpawnEntry() { PrefabName = "Greydwarf",        MaxSpawned = 20, SpawnInterval = 3f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3},
+                        new RaidSpawnEntry() { PrefabName = "Greydwarf_Elite",  MaxSpawned = 6, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.AgitatedByBuild },
+                        new RaidSpawnEntry() { PrefabName = "Greydwarf_Shaman", MaxSpawned = 4, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 5, CreatureAI = AI.Alerted },
                     },
                 }},
                 { new RaidDefinition() {
@@ -101,8 +101,8 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_bonemass" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Skeleton",        MaxSpawned = 6, SpawnInterval = 4f, SpawnChance = 100f, UseRaidLevelSystem = false, CreatureAI = AI.Alerted, SpawnGroupSize = 3 },
-                        new RaidSpawnEntry() { PrefabName = "Skeleton_Poison", MaxSpawned = 2, SpawnInterval = 6f, SpawnChance = 100f, UseRaidLevelSystem = false, CreatureAI = AI.Alerted },
+                        new RaidSpawnEntry() { PrefabName = "Skeleton",        MaxSpawned = 20, SpawnInterval = 4f, SpawnChance = 100f, UseRaidLevelSystem = false, CreatureAI = AI.Alerted, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Skeleton_Poison", MaxSpawned = 8, SpawnInterval = 6f, SpawnChance = 100f, UseRaidLevelSystem = false, CreatureAI = AI.Alerted },
                     },
                 }},
                 { new RaidDefinition() {
@@ -119,8 +119,8 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_bonemass" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Blob",      MaxSpawned = 4, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2 },
-                        new RaidSpawnEntry() { PrefabName = "BlobElite", MaxSpawned = 2, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Blob",      MaxSpawned = 20, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2 },
+                        new RaidSpawnEntry() { PrefabName = "BlobElite", MaxSpawned = 8, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -135,8 +135,8 @@ namespace StarLevelSystem.Data
                         RequiredGlobalKeys = new List<string>() { "defeated_bonemass" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Ghost", MaxSpawned = 3, SpawnInterval = 30f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.AgitatedByBuild },
-                        new RaidSpawnEntry() { PrefabName = "Wraith", MaxSpawned = 3, SpawnInterval = 30f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.AgitatedByBuild },
+                        new RaidSpawnEntry() { PrefabName = "Ghost", MaxSpawned = 12, SpawnInterval = 30f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.AgitatedByBuild },
+                        new RaidSpawnEntry() { PrefabName = "Wraith", MaxSpawned = 12, SpawnInterval = 30f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.AgitatedByBuild },
                     },
                 }},
                 { new RaidDefinition() {
@@ -153,7 +153,7 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_bonemass" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Surtling", MaxSpawned = 4, SpawnInterval = 4f, SpawnChance = 100f, UseRaidLevelSystem = false, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2,
+                        new RaidSpawnEntry() { PrefabName = "Surtling", MaxSpawned = 40, SpawnInterval = 4f, SpawnChance = 100f, UseRaidLevelSystem = false, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2,
                             RequiredModifiers = new Dictionary<string, ModifierType>() { { "Fire", ModifierType.Major } }
                         },
                     },
@@ -172,11 +172,11 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_dragon" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Skeleton",      MaxSpawned = 4, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.Alerted, SpawnGroupSize = 2 },
-                        new RaidSpawnEntry() { PrefabName = "Blob",          MaxSpawned = 2, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.Alerted },
-                        new RaidSpawnEntry() { PrefabName = "Draugr",        MaxSpawned = 3, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.AgitatedByBuild, SpawnGroupSize = 2 },
-                        new RaidSpawnEntry() { PrefabName = "Draugr_Elite",  MaxSpawned = 1, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "Draugr_Ranged", MaxSpawned = 2, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Skeleton",      MaxSpawned = 30, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.Alerted, SpawnGroupSize = 2 },
+                        new RaidSpawnEntry() { PrefabName = "Blob",          MaxSpawned = 10, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.Alerted },
+                        new RaidSpawnEntry() { PrefabName = "Draugr",        MaxSpawned = 12, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.AgitatedByBuild, SpawnGroupSize = 2 },
+                        new RaidSpawnEntry() { PrefabName = "Draugr_Elite",  MaxSpawned = 4, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Draugr_Ranged", MaxSpawned = 8, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 12, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -192,8 +192,8 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_dragon" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Wolf",    MaxSpawned = 5, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
-                        new RaidSpawnEntry() { PrefabName = "Fenring", MaxSpawned = 1, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Wolf",    MaxSpawned = 20, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Fenring", MaxSpawned = 8, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -210,8 +210,8 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_queen" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Ulv",    MaxSpawned = 5, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
-                        new RaidSpawnEntry() { PrefabName = "Fenring_Cultist", MaxSpawned = 2, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Ulv",    MaxSpawned = 20, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Fenring_Cultist", MaxSpawned = 6, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -228,9 +228,9 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_goblinking" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Hatchling",        MaxSpawned = 3, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2 },
-                        new RaidSpawnEntry() { PrefabName = "Wolf",             MaxSpawned = 3, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
-                        new RaidSpawnEntry() { PrefabName = "Fenring_Cultist",  MaxSpawned = 1, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.AgitatedByBuild },
+                        new RaidSpawnEntry() { PrefabName = "Hatchling",        MaxSpawned = 12, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2 },
+                        new RaidSpawnEntry() { PrefabName = "Wolf",             MaxSpawned = 12, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Fenring_Cultist",  MaxSpawned = 6, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 16, CreatureAI = AI.AgitatedByBuild },
                     },
                 }},
                 { new RaidDefinition() {
@@ -247,10 +247,10 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_queen" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Goblin",        MaxSpawned = 4, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2 },
-                        new RaidSpawnEntry() { PrefabName = "GoblinArcher",  MaxSpawned = 2, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "GoblinShaman",  MaxSpawned = 1, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.Alerted },
-                        new RaidSpawnEntry() { PrefabName = "GoblinBrute",   MaxSpawned = 1, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.AgitatedByBuild },
+                        new RaidSpawnEntry() { PrefabName = "Goblin",        MaxSpawned = 16, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2 },
+                        new RaidSpawnEntry() { PrefabName = "GoblinArcher",  MaxSpawned = 12, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "GoblinShaman",  MaxSpawned = 4, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.Alerted },
+                        new RaidSpawnEntry() { PrefabName = "GoblinBrute",   MaxSpawned = 2, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 20, CreatureAI = AI.AgitatedByBuild },
                     },
                 }},
                 { new RaidDefinition() {
@@ -266,7 +266,7 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_queen" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Bat", MaxSpawned = 12, SpawnInterval = 3f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Bat", MaxSpawned = 20, SpawnInterval = 3f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
                     },
                 }},
                 { new RaidDefinition() {
@@ -283,8 +283,8 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_fader" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Gjall",  MaxSpawned = 2, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.AgitatedByBuild },
-                        new RaidSpawnEntry() { PrefabName = "Tick",   MaxSpawned = 8, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2, ModifiersNotAllowed = new List<string>() { "FireNova" } },
+                        new RaidSpawnEntry() { PrefabName = "Gjall",  MaxSpawned = 4, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.AgitatedByBuild },
+                        new RaidSpawnEntry() { PrefabName = "Tick",   MaxSpawned = 16, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 2, ModifiersNotAllowed = new List<string>() { "FireNova" } },
                     },
                 }},
                 { new RaidDefinition() {
@@ -301,9 +301,9 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_fader" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Seeker",       MaxSpawned = 4, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "SeekerBrute",  MaxSpawned = 1, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.AgitatedByBuild },
-                        new RaidSpawnEntry() { PrefabName = "Tick",         MaxSpawned = 2, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Seeker",       MaxSpawned = 12, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "SeekerBrute",  MaxSpawned = 2, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.AgitatedByBuild },
+                        new RaidSpawnEntry() { PrefabName = "Tick",         MaxSpawned = 6, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 26, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -320,9 +320,9 @@ namespace StarLevelSystem.Data
                         NotRequiredGlobalKeys = new List<string>() { "defeated_fader" },
                     },
                     Spawns = new List<RaidSpawnEntry>() {
-                        new RaidSpawnEntry() { PrefabName = "Charred_Twitcher", MaxSpawned = 6, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
-                        new RaidSpawnEntry() { PrefabName = "Charred_Archer",   MaxSpawned = 2, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "Charred_Melee",    MaxSpawned = 2, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Charred_Twitcher", MaxSpawned = 16, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Charred_Archer",   MaxSpawned = 6, SpawnInterval = 12f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Charred_Melee",    MaxSpawned = 4, SpawnInterval = 8f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -355,8 +355,8 @@ namespace StarLevelSystem.Data
                     },
                     Spawns = new List<RaidSpawnEntry>() {
                         new RaidSpawnEntry() { PrefabName = "Skeleton_Hildir",  MaxSpawned = 1, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "Skeleton", MaxSpawned = 8, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
-                        new RaidSpawnEntry() { PrefabName = "Skeleton_Poison",   MaxSpawned = 4, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Skeleton", MaxSpawned = 16, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Skeleton_Poison",   MaxSpawned = 6, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -372,8 +372,8 @@ namespace StarLevelSystem.Data
                     },
                     Spawns = new List<RaidSpawnEntry>() {
                         new RaidSpawnEntry() { PrefabName = "Fenring_Cultist_Hildir",  MaxSpawned = 1, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "Ulv", MaxSpawned = 6, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
-                        new RaidSpawnEntry() { PrefabName = "Fenring_Cultist",   MaxSpawned = 2, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
+                        new RaidSpawnEntry() { PrefabName = "Ulv", MaxSpawned = 16, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Fenring_Cultist", MaxSpawned = 4, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
                     },
                 }},
                 { new RaidDefinition() {
@@ -389,7 +389,7 @@ namespace StarLevelSystem.Data
                     },
                     Spawns = new List<RaidSpawnEntry>() {
                         new RaidSpawnEntry() { PrefabName = "GoblinBruteBros",  MaxSpawned = 1, SpawnInterval = 4f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
-                        new RaidSpawnEntry() { PrefabName = "Goblin", MaxSpawned = 8, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
+                        new RaidSpawnEntry() { PrefabName = "Goblin", MaxSpawned = 16, SpawnInterval = 6f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer, SpawnGroupSize = 3 },
                         new RaidSpawnEntry() { PrefabName = "GoblinShaman",   MaxSpawned = 2, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 1, LevelMax = 30, CreatureAI = AI.HuntPlayer },
                     },
                 }},
@@ -409,7 +409,7 @@ namespace StarLevelSystem.Data
                         new RaidSpawnEntry() { PrefabName = "GoblinBruteBros",  MaxSpawned = 1, SpawnInterval = 180f, SpawnChance = 100f, LevelMin = 15, LevelMax = 30, CreatureAI = AI.HuntPlayer, Faction = Character.Faction.Demon },
                         new RaidSpawnEntry() { PrefabName = "Fenring_Cultist_Hildir",  MaxSpawned = 1, SpawnInterval = 180f, SpawnChance = 100f, LevelMin = 20, LevelMax = 30, CreatureAI = AI.HuntPlayer, Faction = Character.Faction.Demon },
                         new RaidSpawnEntry() { PrefabName = "Skeleton_Hildir",  MaxSpawned = 1, SpawnInterval = 180f, SpawnChance = 100f, LevelMin = 25, LevelMax = 30, CreatureAI = AI.HuntPlayer, Faction = Character.Faction.Demon },
-                        new RaidSpawnEntry() { PrefabName = "Skeleton_Poison",   MaxSpawned = 4, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 20, LevelMax = 30, CreatureAI = AI.HuntPlayer, Faction = Character.Faction.Demon },
+                        new RaidSpawnEntry() { PrefabName = "Skeleton_Poison",   MaxSpawned = 8, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 20, LevelMax = 30, CreatureAI = AI.HuntPlayer, Faction = Character.Faction.Demon },
                         new RaidSpawnEntry() { PrefabName = "GoblinShaman",   MaxSpawned = 2, SpawnInterval = 10f, SpawnChance = 100f, LevelMin = 12, LevelMax = 30, CreatureAI = AI.HuntPlayer, Faction = Character.Faction.Demon },
                     },
                 }},